Class Cil.descriptiveCilPrinterClass

Like defaultCilPrinterClass, but instead of temporary variable names it prints the description that was provided when the temp was created. This is usually better for messages that are printed for end users, although you may want the temporary names for debugging.

The boolean here enables descriptive printing. Usually use true here, but you can set enable to false to make this class behave like defaultCilPrinterClass. This allows subclasses to turn the feature off.

inherit cilPrinter

method startTemps : unit -> unit

method stopTemps : unit -> unit

method pTemps : unit -> GoblintCil__.Pretty.doc