Module Setp.Make

Functor building an implementation of the set structure given a totally ordered type.

Parameters

module Ord : PolyOrderedType

Signature

type 'a elt = 'a Ord.t

The type of the set elements.

type 'a t

The type of sets.

val empty : 'a t

The empty set.

val is_empty : 'a t -> bool

Test whether a set is empty or not.

val mem : 'a elt -> 'a t -> bool

mem x s tests whether x belongs to the set s.

val add : 'a elt -> 'a t -> 'a t

add x s returns a set containing all elements of s, plus x. If x was already in s, s is returned unchanged.

val singleton : 'a elt -> 'a t

singleton x returns the one-element set containing only x.

val remove : 'a elt -> 'a t -> 'a t

remove x s returns a set containing all elements of s, except x. If x was not in s, s is returned unchanged.

val union : 'a t -> 'a t -> 'a t

Set union.

val inter : 'a t -> 'a t -> 'a t

Set interseection.

val diff : 'a t -> 'a t -> 'a t

Set difference.

val compare : 'a t -> 'a t -> int

Total ordering between sets. Can be used as the ordering function for doing sets of sets.

val equal : 'a t -> 'a t -> bool

equal s1 s2 tests whether the sets s1 and s2 are equal, that is, contain equal elements.

val subset : 'a t -> 'a t -> bool

subset s1 s2 tests whether the set s1 is a subset of the set s2.

val iter : ( 'a elt -> unit ) -> 'a t -> unit

iter f s applies f in turn to all elements of s. The order in which the elements of s are presented to f is unspecified.

val fold : ( 'a elt -> 'b -> 'b ) -> 'a t -> 'b -> 'b

fold f s a computes (f xN ... (f x2 (f x1 a))...), where x1 ... xN are the elements of s. The order in which elements of s are presented to f is unspecified.

val for_all : ( 'a elt -> bool ) -> 'a t -> bool

for_all p s checks if all elements of the set satisfy the predicate p.

val exists : ( 'a elt -> bool ) -> 'a t -> bool

exists p s checks if at least one element of the set satisfies the predicate p.

val filter : ( 'a elt -> bool ) -> 'a t -> 'a t

filter p s returns the set of all elements in s that satisfy predicate p.

val partition : ( 'a elt -> bool ) -> 'a t -> 'a t * 'a t

partition p s returns a pair of sets (s1, s2), where s1 is the set of all the elements of s that satisfy the predicate p, and s2 is the set of all the elements of s that do not satisfy p.

val cardinal : 'a t -> int

Return the number of elements of a set.

val elements : 'a t -> 'a elt list

Return the list of all elements of the given set. The returned list is sorted in increasing order with respect to the ordering Ord.compare, where Ord is the argument given to Set.Make.

val min_elt : 'a t -> 'a elt

Return the smallest element of the given set (with respect to the Ord.compare ordering), or raise Not_found if the set is empty.

val max_elt : 'a t -> 'a elt

Same as Set.S.min_elt, but returns the largest element of the given set.

val choose : 'a t -> 'a elt

Return one element of the given set, or raise Not_found if the set is empty. Which element is chosen is unspecified, but equal elements will be chosen for equal sets.