Class FuncFunction.find_calls_with_tmp

method vvdec : GoblintCil.varinfo -> GoblintCil.varinfo GoblintCil.visitAction

Invoked for each variable declaration. The subtrees to be traversed are those corresponding to the type and attributes of the variable. Note that variable declarations are all the GVar, GVarDecl, GFun, all the varinfo in formals of function types, and the formals and locals for function definitions. This means that the list of formals in a function definition will be traversed twice, once as part of the function type and second as part of the formals in a function definition.

method vvrbl : GoblintCil.varinfo -> GoblintCil.varinfo GoblintCil.visitAction

Invoked on each variable use. Here only the SkipChildren and ChangeTo actions make sense since there are no subtrees. Note that the type and attributes of the variable are not traversed for a variable use

method vexpr : GoblintCil.exp -> GoblintCil.exp GoblintCil.visitAction

Invoked on each expression occurrence. The subtrees are the subexpressions, the types (for a Cast or SizeOf expression) or the variable use.

method vlval : GoblintCil.lval -> GoblintCil.lval GoblintCil.visitAction

Invoked on each lvalue occurrence

method voffs : GoblintCil.offset -> GoblintCil.offset GoblintCil.visitAction

Invoked on each offset occurrence that is *not* as part of an initializer list specification, i.e. in an lval or recursively inside an offset.

method vinitoffs : GoblintCil.offset -> GoblintCil.offset GoblintCil.visitAction

Invoked on each offset appearing in the list of a CompoundInit initializer.

method vinst : GoblintCil.instr -> GoblintCil.instr list GoblintCil.visitAction

Invoked on each instruction occurrence. The ChangeTo action can replace this instruction with a list of instructions

method vstmt : GoblintCil.stmt -> GoblintCil.stmt GoblintCil.visitAction

Control-flow statement. The default DoChildren action does not create a new statement when the components change. Instead it updates the contents of the original statement. This is done to preserve the sharing with Goto and Case statements that point to the original statement. If you use the ChangeTo action then you should take care of preserving that sharing yourself.

method vblock : GoblintCil.block -> GoblintCil.block GoblintCil.visitAction

Block.

method vfunc : GoblintCil.fundec -> GoblintCil.fundec GoblintCil.visitAction

Function definition. Replaced in place.

method vglob : GoblintCil.global -> GoblintCil.global list GoblintCil.visitAction

Global (vars, types, etc.)

method vinit : GoblintCil.varinfo -> GoblintCil.offset -> GoblintCil.init -> GoblintCil.init GoblintCil.visitAction

Initializers for static, const and global variables, pass the variable where this occurs, and the offset

method vtype : GoblintCil.typ -> GoblintCil.typ GoblintCil.visitAction

Use of some type. Note that for structure/union and enumeration types the definition of the composite type is not visited. Use vglob to visit it.

method vattr : GoblintCil.attribute -> GoblintCil.attribute list GoblintCil.visitAction

Attribute. Each attribute can be replaced by a list

method vattrparam : GoblintCil.attrparam -> GoblintCil.attrparam GoblintCil.visitAction

Attribute parameters.

method queueInstr : GoblintCil.instr list -> unit

Add here instructions while visiting to queue them to precede the current statement or instruction being processed. Use this method only when you are visiting an expression that is inside a function body, or a statement, because otherwise there will no place for the visitor to place your instructions.

method unqueueInstr : unit -> GoblintCil.instr list

Gets the queue of instructions and resets the queue. This is done automatically for you when you visit statements.