goblint-cil index

Library goblint-cil

This library exposes the following toplevel modules:

Alpha ALPHA conversion

Bitmap

Cabs This file was originally part of Hugues Casee's frontc 2.0, and has been * extensively changed since. ** ** 1.0 3.22.99 Hugues Cassé First version. ** 2.0 George Necula 12/12/00: Many extensions *

Cabs2cil

Cabshelper

Cabsvisit

Cfg Code to compute the control-flow graph of a function or file. This will fill in the preds and succs fields of Cil.stmt

Check

Cil CIL API Documentation.

Cilint

Cillower A number of lowering passes over CIL

Ciloptions

Ciltools

Cilutil

Cilversion

Clexer

Clist Utilities for managing "concatenable lists" (clists). We often need to concatenate sequences, and using lists for this purpose is expensive. This module provides routines to manage such lists more efficiently. In this model, we never do cons or append explicitly. Instead we maintain the elements of the list in a special data structure. Routines are provided to convert to/from ordinary lists, and carry out common list operations.

Cparser

Cprint

Dataflow A framework for data flow analysis for CIL code. Before using this framework, you must initialize the Control-flow Graph for your program, e.g using Cfg.computeFileCFG

Dominators Compute dominators using data flow analysis

Errormsg Utility functions for error-reporting

Escape

Expcompare

Feature Extending CIL with external features

Formatcil An Interpreter for constructing CIL constructs

Formatlex

Formatparse

Frontc

GrowArray Array operations.

Inthash

Lexerhack

Libmaincil

Longarray

Machdep

Machdepenv

Mergecil

Patch

Pretty Utility functions for pretty-printing. The major features provided by this module are

Rmtmps

Stats Utilities for maintaining timing statistics

Trace

Util A bunch of generally useful functions

Whitetrack

Library goblint-cil.dataslicing

The entry point of this library is the module: Dataslicing.

Library goblint-cil.liveness

This library exposes the following toplevel modules:

Liveness

Usedef

Library goblint-cil.makecfg

The entry point of this library is the module: MakeCFG.

Library goblint-cil.pta

This library exposes the following toplevel modules:

Golf

Olf

Ptranal

Setp Sets over ordered types.

Steensgaard

Uref

Library goblint-cil.syntacticsearch

This library exposes the following toplevel modules:

CodeQuery

FuncDatatype

FuncFunction

FuncVar

QueryMapping

ResultPrinter

Library goblint-cil.zrapp

This library exposes the following toplevel modules:

Availexps

Availexpslv

Deadcodeelim

Reachingdefs

Rmciltmps

Zrapp