Module Dataflow.ForwardsDataFlow

Parameters

module T : ForwardsTransfer

Signature

val compute : Cil.stmt list -> unit

Fill in the T.stmtStartData, given a number of initial statements to start from. All of the initial statements must have some entry in T.stmtStartData (i.e., the initial data should not be bottom)