Module GoblintCil.Escape

type wchar = int64

OCaml types used to represent wide characters and strings

type wstring = wchar list

val escape_char : char -> string

escape various constructs in accordance with C lexical rules

val escape_string : string -> string

val escape_wchar : wchar -> string

val escape_wstring : wstring -> string