Module GoblintCil.Mergecil

val ignore_merge_conflicts : bool Stdlib.ref

Set this to true to ignore the merge conflicts

val merge_inlines : bool Stdlib.ref

Try to merge definitions of inline functions. They can appear in multiple files and we would like them all to be the same. This can slow down the merger an order of magnitude !!!

val merge : Cil.file list -> string -> Cil.file

Merge a number of CIL files